implements domain notification pattern

+++ OrdemServico/Incluir/Domain/Impostos.cs
@@ -4,10 +4,10 @@ namespace omie_poc.OrdemServico
     {
         public string cRetemIRRF { get; set; }
         public string cRetemPIS { get; set; }
-        public int nAliqCOFINS { get; set; }
-        public int nAliqCSLL { get; set; }
-        public int nAliqIRRF { get; set; }
-        public int nAliqISS { get; set; }
+        public double nAliqCOFINS { get; set; }
+        public double nAliqCSLL { get; set; }
+        public double nAliqIRRF { get; set; }
+        public double nAliqISS { get; set; }
         public double nAliqPIS { get; set; }
     }
 }
